[GCC] Deal with GCC 64-bit typecasting issues

+++ src/core/lib_objects.cpp
@@ -185,7 +185,7 @@ static ERR object_free(Object *Object)
       if (auto lock = std::unique_lock{glmTimer, 200ms}) {
          for (auto it=glTimers.begin(); it != glTimers.end(); ) {
             if (it->SubscriberID IS Object->UID) {
-               log.warning("%s object #%d has an unfreed timer subscription, routine %p, interval %" PF64, mc->ClassName, Object->UID, &it->Routine, it->Interval);
+               log.warning("%s object #%d has an unfreed timer subscription, routine %p, interval %" PF64, mc->ClassName, Object->UID, &it->Routine, (long long)it->Interval);
                it = glTimers.erase(it);
             }
             else it++;
